+++ packages/dso-toolkit/src/components/form/form.args.ts
@@ -5,25 +5,12 @@ import { Form, FormAsteriskExplanationPosition } from "./models/form.model.js";
 
 export interface FormArgs {
   asteriskExplanation?: FormAsteriskExplanationPosition;
-  legend?: string;
-  legendHeading?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | undefined;
   mode: "horizontal" | "vertical" | undefined;
 }
 
 export const formArgTypes: ArgTypes<FormArgs> = {
   asteriskExplanation: {
     options: [undefined, "top", "bottom", "both"],
-    control: {
-      type: "select",
-    },
-  },
-  legend: {
-    control: {
-      type: "text",
-    },
-  },
-  legendHeading: {
-    options: [undefined, "h1", "h2", "h3", "h4", "h5", "h6"],
     control: {
       type: "select",
     },
@@ -39,10 +26,8 @@ export const formArgTypes: ArgTypes<FormArgs> = {
 export function formArgsMapper<TemplateFnReturnType>(a: FormArgs): Form<TemplateFnReturnType> {
   return {
     asteriskExplanation: a.asteriskExplanation,
-    legend: a.legend,
-    legendHeading: a.legendHeading,
     mode: a.mode,
-    formGroups: content,
+    content,
     formButtons: buttons,
   };
 }