+++ packages/dso-toolkit/src/components/form/models/form.model.ts
@@ -3,11 +3,14 @@ import { FormGroup } from "./form-group.model.js";
 
 export interface Form<TemplateFnReturnType> {
   asteriskExplanation?: FormAsteriskExplanationPosition;
-  legend?: string;
-  legendHeading?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | undefined;
   mode?: "vertical" | "horizontal";
-  formGroups: FormGroup<TemplateFnReturnType>[] | TemplateFnReturnType;
+  content: FormGroup<TemplateFnReturnType>[] | FormGroupCollection<TemplateFnReturnType>[];
   formButtons?: FormButtons;
 }
 
 export type FormAsteriskExplanationPosition = "top" | "bottom" | "both";
+
+export interface FormGroupCollection<TemplateFnReturnType> {
+  title: string;
+  formGroups: FormGroup<TemplateFnReturnType>[];
+}