+++ storybook/src/components/form/form.css-template.ts
@@ -1,4 +1,4 @@
-import { Form } from "dso-toolkit";
+import { Form, FormGroupCollection } from "dso-toolkit";
 import { html, nothing, TemplateResult } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 
@@ -8,7 +8,7 @@ export const cssForm: ComponentImplementation<Form<TemplateResult>> = {
   component: "form",
   implementation: "html-css",
   template: ({ formGroupTemplate, formButtonsTemplate }) =>
-    function formTemplate({ asteriskExplanation, legend, legendHeading, mode, formGroups, formButtons }) {
+    function formTemplate({ asteriskExplanation, mode, content, formButtons }) {
       function asteriskExplanationTemplate() {
         return html`<div class="form-explanation" aria-hidden="true">
           <p class="form-explanation-text">
@@ -17,28 +17,20 @@ export const cssForm: ComponentImplementation<Form<TemplateResult>> = {
         </div>`;
       }
 
+      function formGroupCollection({ title, formGroups }: FormGroupCollection<TemplateResult>) {
+        return html`<fieldset class="dso-form-group-collection">
+          <legend>${title}</legend>
+          ${formGroups.map((formGroup) => formGroupTemplate(formGroup))}
+        </fieldset>`;
+      }
+
       return html`
         <form class=${ifDefined(mode === "horizontal" ? "form-horizontal" : undefined)}>
           ${asteriskExplanation === "top" || asteriskExplanation === "both" ? asteriskExplanationTemplate() : nothing}
-          <fieldset>
-            <legend>
-              ${legendHeading === "h1" // if
-                ? html`<h1>${legend}</h1>`
-                : legendHeading === "h2" // else if
-                  ? html`<h2>${legend}</h2>`
-                  : legendHeading === "h3" // else if
-                    ? html`<h3>${legend}</h3>`
-                    : legendHeading === "h4" // else if
-                      ? html`<h4>${legend}</h4>`
-                      : legendHeading === "h5" // else if
-                        ? html`<h5>${legend}</h5>`
-                        : legendHeading === "h6" // else if
-                          ? html`<h6>${legend}</h6>`
-                          : html`${legend}`}
-            </legend>
-            ${"_$litType$" in formGroups ? formGroups : formGroups.map((formGroup) => formGroupTemplate(formGroup))}
-            ${formButtons ? formButtonsTemplate(formButtons) : nothing}
-          </fieldset>
+          ${content.map((formGroup) =>
+            "title" in formGroup ? formGroupCollection(formGroup) : formGroupTemplate(formGroup),
+          )}
+          ${formButtons ? formButtonsTemplate(formButtons) : nothing}
           ${asteriskExplanation === "bottom" || asteriskExplanation === "both"
             ? asteriskExplanationTemplate()
             : nothing}